Create exhaustive_trefoil executable
This program is intended to exhaustively enumerate joint chains to find the "knot" presentations (which are not, at this point, actually guaranteed to have any particular topology) whose minimum costs (or at least local minimums, determined in a way that privileges one chirality over another) are lowest.

Tools for visualizing the output of this executable are forthcoming.