Increase initial radius
Empirical investigation with symmetry_adjust_analysis suggests that an initial radius of 10.0 produces a more stable gradient descent process, although it does sometimes result in the nearest local minima being chosen in a way that seems to select against knots.

The default knot in interactive_symmetry_adjust has been updated to reflect new results from exhaustive_trefoil as a result of this change.